Output functions in their own module, --flat flag.
Output functions live in their own module now, and there is --flat flag
which causes all output to be single line (as opposed to default
'pretty' JSON). This will be combined with --silent or something like
that so that no prompt will be displayed, for batch processing of
requests from stdin.