Added 'edit' command, request history
'edit' command allows you to edit the most recent request using VIM, and
to re-run it. The most recent request is now stored on exit and read on
startup, and when you run 'search' and 'count' without any parameters,
this most recent request gets executed. All this functionality is now
documented and part of the TUTORIAL.md

Version bumped to 0.9.3