Better sanitation of CLI argument 'index'.
Issue #9. Make sure that the 'index' argument to elsec is one of the
existing indexes or aliases. It was possible to pass
'valid_index/valid_doc_type' combnation (this was not a feature) which
lead to all kinds of grief down the road when the mapping was not parsed
properly.

Version bump to 1.0.3